Fixed datatype error
double not int, thats funny.

+++ FRC-Training-2021/src/main/java/frc/robot/commands/ExampleCommand.java
@@ -23,7 +23,7 @@ public class ExampleCommand extends CommandBase {
   private Joystick m_joystick;
   
   // Variables to hold data about joystick positioning
-  private int joy_x, joy_y;
+  private double joy_x, joy_y;
 
   
   public ExampleCommand(ExampleSubsystem subsystem) {
